+++ tests/test_fit_methods.py
@@ -23,8 +23,8 @@ def test_roc_stable_fit(X_y_dates_romania):
     """Only integration test"""
     X, y, dates = X_y_dates_romania
     dates = dates.astype('datetime64[D]').astype('int')
-    beta, resid, is_stable = fm.roc_stable_fit(X.astype(np.float32),
-                                               y.astype(np.float32),
+    beta, resid, is_stable = fm.roc_stable_fit(X.astype(np.float64),
+                                               y.astype(np.float64),
                                                dates)
